Moving collaborator API call out of resources files.
