+++ timesketch/api/v1/resources.py
@@ -3392,70 +3392,6 @@ class GroupListResource(ResourceMixin, Resource):
         return self.to_json(Group.query.all())
 
 
-class CollaboratorResource(ResourceMixin, Resource):
-    """Resource to update sketch collaborators."""
-
-    @login_required
-    def post(self, sketch_id):
-        """Handles POST request to the resource.
-
-        Args:
-            sketch_id: Integer primary key for a sketch database model
-        """
-        sketch = Sketch.query.get_with_acl(sketch_id)
-        if not sketch:
-            abort(
-                HTTP_STATUS_CODE_NOT_FOUND, 'No sketch found with this ID.')
-        form = request.json
-
-        # TODO: Add granular ACL controls.
-        # https://github.com/google/timesketch/issues/1016
-
-        if not sketch.has_permission(user=current_user, permission='write'):
-            abort(
-                HTTP_STATUS_CODE_FORBIDDEN,
-                'The user does not have write permission on the sketch.')
-
-        for username in form.get('users', []):
-            # Try the username with any potential @domain preserved.
-            user = User.query.filter_by(username=username).first()
-
-            # If no hit, then try to strip the domain.
-            if not user:
-                base_username = username.split('@')[0]
-                base_username = base_username.strip()
-                user = User.query.filter_by(username=base_username).first()
-
-            if user:
-                sketch.grant_permission(permission='read', user=user)
-                sketch.grant_permission(permission='write', user=user)
-
-        for group in form.get('groups', []):
-            group = Group.query.filter_by(name=group).first()
-            # Only add groups publicly visible or owned by the current user
-            if not group.user or group.user == current_user:
-                sketch.grant_permission(permission='read', group=group)
-                sketch.grant_permission(permission='write', group=group)
-
-        for username in form.get('remove_users', []):
-            user = User.query.filter_by(username=username).first()
-            sketch.revoke_permission(permission='read', user=user)
-            sketch.revoke_permission(permission='write', user=user)
-
-        for group in form.get('remove_groups', []):
-            group = Group.query.filter_by(name=group).first()
-            sketch.revoke_permission(permission='read', group=group)
-            sketch.revoke_permission(permission='write', group=group)
-
-        public = form.get('public')
-        if public == 'true':
-            sketch.grant_permission(permission='read')
-        else:
-            sketch.revoke_permission(permission='read')
-
-        return HTTP_STATUS_CODE_OK
-
-
 class SessionResource(ResourceMixin, Resource):
     """Resource to get sessions."""
